Move secrets_file fixture to conftest
This fixture is needed by more modules in this package now.
