+++ tests/api/auth_test.py
@@ -55,17 +55,3 @@ def test_read_users_when_given_a_valid_file_then_a_proper_dict_should_be_returne
     for user in ("user1", "user2", "user3"):
         assert user in users
         assert len(users[user]) == 40
-
-
-@pytest.fixture
-def secrets_file(tmp_path):
-    name = tmp_path.joinpath("secrets")
-    with open(name, "w") as conf:
-        conf.write(
-            """user1 3c55d3cdc19876dfc8c0bb49da8c927a0ddff26d
-            user2 eb6fb35f5fbba6a6e43d3c893ad7a77dc793ceba
-
-            user3 c8a0b250edb2eabd9616b7b05a46e0ad28226fc2
-            """
-        )
-    yield name
